Multiple target sizes and test mode
Added the ability to resize to/from multiple sizes within the same cloud service. 
Also added a default "test" mode - that that instead of running the conversion, it creates a PowerShell file of the commands needed to perform the conversion. This allows an SA to review the scripts then run the generated PowerShell file to actually perform the conversion.